+++ cm-mlops/script/get-gcc/customize.py
@@ -22,18 +22,26 @@ def preprocess(i):
                                        'run_script_input':i['run_script_input'],
                                        'recursion_spaces':recursion_spaces})
     if r['return'] >0 : 
-       if r['return'] == 16:
-           if env.get('CM_TMP_FAIL_IF_NOT_FOUND','').lower() == 'yes':
-               return r
-
-           print (recursion_spaces+'    # {}'.format(r['error']))
-
-           # Attempt to run installer
-           r = {'return':0, 'skip':True, 'script':{'tags':'install,prebuilt,gcc'}}
+#       if r['return'] == 16:
+#           if env.get('CM_TMP_FAIL_IF_NOT_FOUND','').lower() == 'yes':
+#               return r
+#
+#           print (recursion_spaces+'    # {}'.format(r['error']))
+#
+#           # Attempt to run installer
+#           r = {'return':0, 'skip':True, 'script':{'tags':'install,prebuilt,gcc'}}
 
        return r
 
-    found_path = r['found_path']
+    # G: changed next line to handle cases like gcc-8
+    # found_path = r['found_path']
+    full_path = r['full_path']
+
+    found_path = os.path.dirname(full_path)
+    found_file = os.path.basename(full_path)
+
+    file_name_c = found_file
+    file_name_cpp = found_file.replace('gcc','g++')
 
     env['CM_GCC_BIN']=file_name_c
     env['CM_GCC_BIN_WITH_PATH']=os.path.join(found_path, file_name_c)
@@ -51,7 +59,7 @@ def preprocess(i):
     env['DEBUG_LINKER_FLAGS'] = "-O0"
     env['DEFAULT_COMPILER_FLAGS'] = "-O2"
     env['DEFAULT_LINKER_FLAGS'] = "-O2"
-    
+
     return {'return':0}
 
 
@@ -59,7 +67,7 @@ def postprocess(i):
 
     env = i['env']
 
-    r = i['automation'].parse_version({'match_text': r'gcc \(.*\)\s*([\d.]+)',
+    r = i['automation'].parse_version({'match_text': r' \(.*\)\s*([\d.]+)',
                                        'group_number': 1,
                                        'env_key':'CM_GCC_VERSION',
                                        'which_env':i['env']})
@@ -69,6 +77,7 @@ def postprocess(i):
     version = r['version']
 
     print (i['recursion_spaces'] + '    Detected version: {}'.format(version))
+
     env['CM_COMPILER_FAMILY'] = 'GCC'
     env['CM_COMPILER_VERSION'] = env['CM_GCC_VERSION']
     env['CM_GCC_CACHE_TAGS'] = 'version-'+version