+++ cm/cmind/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.2"
+__version__ = "1.0.2.1"
 
 from cmind.core import access
 from cmind.core import error